pyproject.toml
[tool.poetry]
name = "tcc-madr"
version = "0.1.0"
description = ""
authors = ["GermanoC <105460981+GermanoCesarSS@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.1"
factory-boy = "^3.3.1"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations', './tests/utils/']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["E117", "PLR0913", "PLR0917"]

[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev tcc_madr/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=tcc_madr -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
